Replace flaky navigation adaptor tests with focused time-range tests
The navigation adaptor tests drove a custom continuation-based stores
manager whose mutable state was touched from two executors without
synchronisation, deadlocking under Swift Testing parallel mode. CI test
36885 was cancelled after a ~10 hour hang on this suite.

The navigation tests themselves were a mix of type-system assertions
("dispatches on the main thread") and UI-flow checks tightly coupled to
the @MainActor Task that powered the deadlock. Drop them rather than
rebuild the mock plumbing - the loader-vs-placeholder behaviour is
better verified end-to-end through /verify, and the adaptor's
non-trivial pure logic is the analytics time-range parser.

Preserve the three time-range tests in a focused new file. They
exercise pure date and timezone parsing logic with a trivial
MockStoresManager so the adaptor can be constructed - no continuations,
no async waits, no flakiness surface.
